continuing

+++ filters/approximatecoplanar/ApproximateCoplanarFilter.cpp
@@ -97,7 +97,7 @@ void ApproximateCoplanarFilter::filter(PointView& view)
         auto B = computeCovariance(view, ids);
 
         // perform the eigen decomposition
-        SelfAdjointEigenSolver<Matrix3f> solver(B);
+        SelfAdjointEigenSolver<Matrix3d> solver(B);
         if (solver.info() != Success)
             throw pdal_error("Cannot perform eigen decomposition.");
         auto ev = solver.eigenvalues();
