+++ filters/eigenvalues/EigenvaluesFilter.cpp
@@ -95,7 +95,7 @@ void EigenvaluesFilter::filter(PointView& view)
         auto B = computeCovariance(view, ids);
 
         // perform the eigen decomposition
-        SelfAdjointEigenSolver<Matrix3f> solver(B);
+        SelfAdjointEigenSolver<Matrix3d> solver(B);
         if (solver.info() != Success)
             throw pdal_error("Cannot perform eigen decomposition.");
         auto ev = solver.eigenvalues();
